:bug: fixed models import and renamed to backbone instead of encoder
